Renamed utils package to commands.
