+++ commands/commands_test.go
@@ -1,4 +1,4 @@
-package utils
+package commands
 
 import (
 	"io/ioutil"
